# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Teleasistencia Documentation",
        "version": "1.0.0"
    },
    "paths": {
...
        "/api/reports/check-emergencies": {
            "get": {
                "tags": [
                    "Reports"
                ],
                "summary": "Verificar emergencias en una zona",
                "operationId": "25046cd5749b2fa196406c8c6faf35d3",
                "parameters": [
                    {
                        "name": "zone_id",
                        "in": "query",
                        "description": "ID de la zona",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    },
                    {
                        "name": "start_date",
                        "in": "query",
                        "description": "Fecha de inicio (Y-m-d)",
                        "required": false,
                        "schema": {
                            "type": "string",
                            "format": "date"
                        }
                    },
                    {
                        "name": "end_date",
                        "in": "query",
                        "description": "Fecha de fin (Y-m-d)",
                        "required": false,
                        "schema": {
                            "type": "string",
                            "format": "date"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Respuesta de verificación de emergencias",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/reports/emergency-report": {
            "get": {
                "tags": [
                    "Reports"
                ],
                "summary": "Generar informe de emergencias para una zona",
                "operationId": "780ba93f2b26e633fde3187cee0bc9f2",
                "parameters": [
                    {
                        "name": "zone_id",
                        "in": "query",
                        "description": "ID de la zona",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    },
                    {
                        "name": "start_date",
                        "in": "query",
                        "description": "Fecha de inicio (Y-m-d)",
                        "required": false,
                        "schema": {
                            "type": "string",
                            "format": "date"
                        }
                    },
                    {
                        "name": "end_date",
                        "in": "query",
                        "description": "Fecha de fin (Y-m-d)",
                        "required": false,
                        "schema": {
                            "type": "string",
                            "format": "date"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "PDF del informe de emergencias"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/reports/pacientes": {
            "get": {
                "tags": [
                    "Reports"
                ],
                "summary": "Obtener lista de pacientes ordenados por apellido",
                "operationId": "655c3acd2ece54a646ca22b62d41f8d0",
                "responses": {
                    "200": {
                        "description": "Lista de pacientes",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/PatientResource"
                                    }
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/reports/llamadas-programadas": {
            "get": {
                "tags": [
                    "Reports"
                ],
                "summary": "Obtener llamadas programadas",
                "operationId": "78fd75f7c9e58d9531ca8b45740e7333",
                "parameters": [
                    {
                        "name": "format",
                        "in": "query",
                        "description": "Formato de exportación (pdf/csv)",
                        "required": false,
                        "schema": {
                            "type": "string",
                            "enum": [
                                "pdf",
                                "csv"
                            ]
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Listado de llamadas programadas"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/reports/llamadas-realizadas": {
            "get": {
                "tags": [
                    "Reports"
                ],
                "summary": "Obtener llamadas completadas del día",
                "operationId": "55677ad4957570cb5909fae5fd0e1b97",
                "parameters": [
                    {
                        "name": "date",
                        "in": "query",
                        "description": "Fecha a consultar (Y-m-d)",
                        "required": false,
                        "schema": {
                            "type": "string",
                            "format": "date"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Listado de llamadas completadas"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/reports/historial-paciente/{id}": {
            "get": {
                "tags": [
                    "Reports"
                ],
                "summary": "Obtener historial de llamadas de un paciente",
                "operationId": "2dd71a50ead1927aebe0c8471a5ad7ff",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID del paciente",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    },
                    {
                        "name": "start_date",
                        "in": "query",
                        "description": "Fecha de inicio (Y-m-d)",
                        "required": false,
                        "schema": {
                            "type": "string",
                            "format": "date"
                        }
                    },
                    {
                        "name": "end_date",
                        "in": "query",
                        "description": "Fecha de fin (Y-m-d)",
                        "required": false,
                        "schema": {
                            "type": "string",
                            "format": "date"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Historial de llamadas del paciente"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "schemas": {
...
            "PatientResource": {
                "description": "Esquema del recurso Paciente",
                "properties": {
                    "id": {
                        "description": "Identificador único del paciente",
                        "type": "integer",
                        "example": 1
                    },
                    "nombre": {
                        "description": "Nombre completo del paciente",
                        "type": "string",
                        "example": "María López"
                    },
                    "fecha_nacimiento": {
                        "description": "Fecha de nacimiento del paciente",
                        "type": "string",
                        "format": "date",
                        "example": "1985-04-12"
                    },
                    "direccion": {
                        "description": "Dirección del paciente",
                        "type": "string",
                        "example": "Calle Falsa 123, Madrid"
                    },
                    "ciudad": {
                        "description": "Ciudad de residencia del paciente",
                        "type": "string",
                        "example": "Madrid",
                        "nullable": true
                    },
                    "codigo_postal": {
                        "description": "Código postal del paciente",
                        "type": "string",
                        "maxLength": 10,
                        "example": "28080",
                        "nullable": true
                    },
                    "dni": {
                        "description": "DNI único del paciente",
                        "type": "string",
                        "maxLength": 20,
                        "example": "12345678A"
                    },
                    "tarjeta_sanitaria": {
                        "description": "Número de tarjeta sanitaria único",
                        "type": "string",
                        "maxLength": 50,
                        "example": "TS1234567890"
                    },
                    "telefono": {
                        "description": "Número de teléfono del paciente",
                        "type": "string",
                        "maxLength": 20,
                        "example": "+34 612 345 678"
                    },
                    "email": {
                        "description": "Correo electrónico del paciente",
                        "type": "string",
                        "format": "email",
                        "maxLength": 100,
                        "example": "maria.lopez@example.com",
                        "nullable": true
                    },
                    "zona_id": {
                        "description": "ID de la zona asignada al paciente",
                        "type": "integer",
                        "example": 3
                    },
                    "situacion_personal": {
                        "description": "Situación personal del paciente",
                        "type": "string",
                        "example": "Vive solo",
                        "nullable": true
                    },
                    "estado_salud": {
                        "description": "Estado de salud actual del paciente",
                        "type": "string",
                        "example": "Hipertensión",
                        "nullable": true
                    },
                    "condicion_vivienda": {
                        "description": "Condición de la vivienda del paciente",
                        "type": "string",
                        "example": "Vivienda propia en buen estado",
                        "nullable": true
                    },
                    "nivel_autonomia": {
                        "description": "Nivel de autonomía del paciente",
                        "type": "string",
                        "example": "Autónomo",
                        "nullable": true
                    },
                    "situacion_economica": {
                        "description": "Situación económica del paciente",
                        "type": "string",
                        "example": "Media-baja",
                        "nullable": true
                    },
                    "created_at": {
                        "description": "Fecha de creación del paciente",
                        "type": "string",
                        "format": "date-time",
                        "example": "2025-02-23 12:00:00"
                    },
                    "updated_at": {
                        "description": "Fecha de última actualización del paciente",
                        "type": "string",
                        "format": "date-time",
                        "example": "2025-02-24 12:00:00"
                    }
                },
                "type": "object"
            },
...
        },
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
...
}